Fix add limits and regularization for Q estimates. Remove data downsizing in replay buffer.

+++ utils/replay_buffer.py
@@ -1,6 +1,7 @@
 import numpy as np
 import random
 
+
 def sample_n_unique(sampling_f, n):
     """Helper function. Given a function `sampling_f` that returns
     comparable objects, sample n such unique objects.
@@ -12,10 +13,12 @@ def sample_n_unique(sampling_f, n):
             res.append(candidate)
     return res
 
+
 class ReplayBuffer(object):
     """
     Taken from Berkeley's Assignment
     """
+
     def __init__(self, size, frame_history_len):
         """This is a memory efficient implementation of the replay buffer.
 
@@ -45,27 +48,30 @@ class ReplayBuffer(object):
         self.size = size
         self.frame_history_len = frame_history_len
 
-        self.next_idx      = 0
+        self.next_idx = 0
         self.num_in_buffer = 0
 
-        self.obs      = None
-        self.action   = None
-        self.reward   = None
-        self.done     = None
+        self.obs = None
+        self.action = None
+        self.reward = None
+        self.done = None
 
     def can_sample(self, batch_size):
         """Returns true if `batch_size` different transitions can be sampled from the buffer."""
         return batch_size + 1 <= self.num_in_buffer
 
     def _encode_sample(self, idxes):
-        obs_batch      = np.concatenate([self._encode_observation(idx)[None] for idx in idxes], 0)
-        act_batch      = self.action[idxes]
-        rew_batch      = self.reward[idxes]
-        next_obs_batch = np.concatenate([self._encode_observation(idx + 1)[None] for idx in idxes], 0)
-        done_mask      = np.array([1.0 if self.done[idx] else 0.0 for idx in idxes], dtype=np.float32)
+        obs_batch = np.concatenate(
+            [self._encode_observation(idx)[None] for idx in idxes], 0)
+        act_batch = self.action[idxes]
+        rew_batch = self.reward[idxes]
+        next_obs_batch = np.concatenate(
+            [self._encode_observation(idx + 1)[None] for idx in idxes], 0)
+        done_mask = np.array([1.0 if self.done[idx] else 0.0 for idx in idxes],
+                             dtype=np.float32)
 
+        # print("reward: {}, done: {}".format(rew_batch, done_mask))
         return obs_batch, act_batch, rew_batch, next_obs_batch, done_mask
-
 
     def sample(self, batch_size):
         """Sample `batch_size` different transitions.
@@ -101,7 +107,8 @@ class ReplayBuffer(object):
             Array of shape (batch_size,) and dtype np.float32
         """
         assert self.can_sample(batch_size)
-        idxes = sample_n_unique(lambda: random.randint(0, self.num_in_buffer - 2), batch_size)
+        idxes = sample_n_unique(
+            lambda: random.randint(0, self.num_in_buffer - 2), batch_size)
         return self._encode_sample(idxes)
 
     def encode_recent_observation(self):
@@ -118,7 +125,7 @@ class ReplayBuffer(object):
         return self._encode_observation((self.next_idx - 1) % self.size)
 
     def _encode_observation(self, idx):
-        end_idx   = idx + 1 # make noninclusive
+        end_idx = idx + 1  # make noninclusive
         start_idx = end_idx - self.frame_history_len
         # this checks if we are using low-dimensional observations, such as RAM
         # state, in which case we just directly return the latest RAM.
@@ -134,14 +141,17 @@ class ReplayBuffer(object):
         # if zero padding is needed for missing context
         # or we are on the boundry of the buffer
         if start_idx < 0 or missing_context > 0:
-            frames = [np.zeros_like(self.obs[0]) for _ in range(missing_context)]
+            frames = [
+                np.zeros_like(self.obs[0]) for _ in range(missing_context)
+            ]
             for idx in range(start_idx, end_idx):
                 frames.append(self.obs[idx % self.size])
             return np.concatenate(frames, 2)
         else:
             # this optimization has potential to saves about 30% compute time \o/
             img_h, img_w = self.obs.shape[1], self.obs.shape[2]
-            return self.obs[start_idx:end_idx].transpose(1, 2, 0, 3).reshape(img_h, img_w, -1)
+            return self.obs[start_idx:end_idx].transpose(1, 2, 0, 3).reshape(
+                img_h, img_w, -1)
 
     def store_frame(self, frame):
         """Store a single frame in the buffer at the next available index, overwriting
@@ -159,11 +169,14 @@ class ReplayBuffer(object):
             Index at which the frame is stored. To be used for `store_effect` later.
         """
         if self.obs is None:
-            self.obs      = np.empty([self.size] + list(frame.shape), dtype=np.uint8)
-            self.action   = np.empty([self.size],                     dtype=np.int32)
-            self.reward   = np.empty([self.size],                     dtype=np.float32)
-            self.done     = np.empty([self.size],                     dtype=np.bool)
+            self.obs = np.empty(
+                [self.size] + list(frame.shape), dtype=np.float32)
+            self.action = np.empty([self.size], dtype=np.int32)
+            self.reward = np.empty([self.size], dtype=np.float32)
+            self.done = np.empty([self.size], dtype=np.bool)
         self.obs[self.next_idx] = frame
+        #print("frame {}", format(frame))
+        #print("stored_frame {}", format(self.obs[self.next_idx]))
 
         ret = self.next_idx
         self.next_idx = (self.next_idx + 1) % self.size
@@ -190,5 +203,4 @@ class ReplayBuffer(object):
         """
         self.action[idx] = action
         self.reward[idx] = reward
-        self.done[idx]   = done
-
+        self.done[idx] = done
